feat(economics): add evm_fee_revenue to block header for correct fee distribution
distribute_fees received evm_gas_used (gas units, e.g. 21000) instead of
actual fee revenue (gas_used × effective_gas_price). This meant EVM-only
blocks distributed ~21000 wei instead of ~2.1 TRS per transaction.

- Add evm_fee_revenue: u128 to TorusBlockHeader and canonical encoding
- Compute from receipts at proposal time via compute_fee_revenue()
- Validators verify evm_fee_revenue matches local execution
- distribute_fees now receives correct fee revenue in wei
- Fix stale chain_id assertion in eth_chain_id RPC test (7777→7778)

+++ crates/torus-bridge/src/native_executor.rs
@@ -1148,8 +1148,8 @@ impl NativeExecutor {
     }
 
     /// Distribute fees at end of block.
-    pub fn distribute_fees(ctx: &mut NativeExecContext, total_evm_fees: u64) -> NativeActionResult {
-        let total_fees = U256::from(ctx.total_native_fees + total_evm_fees);
+    pub fn distribute_fees(ctx: &mut NativeExecContext, total_evm_fees: u128) -> NativeActionResult {
+        let total_fees = U256::from(ctx.total_native_fees as u128 + total_evm_fees);
         if total_fees.is_zero() {
             return NativeActionResult::ok("fee_distribution", 0);
         }
